keyboard_UI: fix #33838 'e' keyboard shortcut for anonymous user.
Previously, when in anonymous login, pressing the "view original
message" button showed the original message and allowed copying it.
However, the corresponding keyboard shortcut ('e') did not work as
expected and a login pop-up appeared instead.

Now, by fixing the UI features' permissions for anonymous users, they
are allowed to use the 'e' keyboard shortcut as intended.

Additionally, I added a test to ensure the bug is properly fixed.